Make ton of differences - might start fresh
I built a large chunk of the webpage and learnt
to use the scemantic tags like main, header, nav,
and footer - and section, but that is for use with
blogs and stuff where you have distinct sections
like chapters etc. - still need to do some research
- on that. I also realised why simple code is
important. Rewriting the same thing e.g.
width = 750px is tedious and inefficient. I might
redo the whole thing tomorrow to have it be more
efficient and easier to work with.

It is a lot of fun though. Like a puzzle to be
figured out.